Enemy Navigation Update
Enemies can navigate around new objects.

Shop does not work, because game objects cannot be placed on top of UI elements when canvas render mode is set to world space. So I made a temporary pile of objects at the bottom for demo purposes. Will need to find workaround for the future.

Enemies were being destroyed instantly on spawn, because they fired an egg which immedietly collided with themselves. So, enemy attack module is temporarily disabled.

Zoom module and mouse movement module are disabled temporarily just for testing. Can be enabled again by checking the box on the inspector.